refactor(core): moved dialect-independent storage machinery from db-postgres to core
