+++ packages/core/package.json
@@ -91,6 +91,7 @@
     "dotenv": "^17.4.2",
     "pino": "^10.3.1",
     "sharp": "^0.35.3",
+    "uuid": "^14.0.1",
     "zod": "^4.4.3"
   },
   "devDependencies": {
