+++ packages/core/src/storage/index.ts
@@ -15,3 +15,11 @@ export {
   fieldTypeToStoreType,
   type StoreType,
 } from './field-store-map.js'
+export { flattenFieldSetData } from './storage-flatten.js'
+export {
+  extractFlattenedFieldValue,
+  type RestoreResult,
+  restoreFieldSetData,
+} from './storage-restore.js'
+export { resolveStoreTypes } from './storage-utils.js'
+export type { FlattenedFieldValue, UnifiedFieldValue } from './storage-row-types.js'
